wlan: embed cld80211 library into legacy qcom wifi HAL
The library is not correctly installed in the vendor partition, causing
the Wi-Fi HAL to fail at runtime. Since the library consists of a single
C file, embed its functionality directly into the QCOM Wi-Fi HAL library.

While on it, comment out a wifi logger error line to avoid spamming the
logcat. The old qcacld-2.0 driver does not support some commands the wifi
HAL is trying to execute (wifi stats, etc), but that does not prevent
wifi from working properly.

https://onedigi.atlassian.net/browse/ADK4A-1940

Change-Id: I8b362bcf88b8d9d543798439aa78125c7ef847dd
Signed-off-by: Javier Viguera <javier.viguera@digi.com>